Rename detector cover for activity to follow naming system on rest of cover detectors
