+++ src/main/java/gregtech/common/covers/detector/CoverDetectorActivity.java
@@ -10,18 +10,23 @@ import gregtech.api.cover.ICoverable;
 import gregtech.client.renderer.texture.Textures;
 import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.EnumFacing;
+import net.minecraft.util.ITickable;
 
-public class CoverActivityDetectorAdvanced extends CoverActivityDetector {
-
-    public CoverActivityDetectorAdvanced(ICoverable coverHolder, EnumFacing attachedSide) {
+public class CoverDetectorActivity extends CoverDetectorBase implements ITickable {
+    public CoverDetectorActivity(ICoverable coverHolder, EnumFacing attachedSide) {
         super(coverHolder, attachedSide);
     }
 
     @Override
-    public void renderCover(CCRenderState renderState, Matrix4 translation, IVertexOperation[] pipeline, Cuboid6 plateBox, BlockRenderLayer layer) {
-        Textures.DETECTOR_ACTIVITY_ADVANCED.renderSided(attachedSide, plateBox, renderState, pipeline, translation);
+    public boolean canAttach() {
+        return coverHolder.getCapability(GregtechTileCapabilities.CAPABILITY_WORKABLE, null) != null;
     }
-    
+
+    @Override
+    public void renderCover(CCRenderState renderState, Matrix4 translation, IVertexOperation[] pipeline, Cuboid6 plateBox, BlockRenderLayer layer) {
+        Textures.DETECTOR_ACTIVITY.renderSided(attachedSide, plateBox, renderState, pipeline, translation);
+    }
+
     @Override
     public void update() {
         if (this.coverHolder.getOffsetTimer() % 20 != 0)
@@ -31,13 +36,7 @@ public class CoverActivityDetectorAdvanced extends CoverActivityDetector {
         if (workable == null)
             return;
 
-        int outputAmount = (int) (15.0 * workable.getProgress() / workable.getMaxProgress());
-
-        if (!workable.isWorkingEnabled())
-            outputAmount = 0;
-        else if (this.isInverted())
-            outputAmount = 15 - outputAmount;
-
-        setRedstoneSignalOutput(outputAmount);
+        if (isInverted()) setRedstoneSignalOutput(workable.isActive() && workable.isWorkingEnabled() ? 0 : 15);
+        else setRedstoneSignalOutput(workable.isActive() && workable.isWorkingEnabled() ? 15 : 0);
     }
 }
